feat(pwa) use offline-plugin for service worker setup and offline sup… (#3000)
* feat(pwa) use offline-plugin for service worker setup and offline support

* feat(pwa) the whole website works offline

* feat(pwa) make sure prod build runs after ssg. Provide css from ssg run to prod SW

* feat(pwa) remove appShell option from offline-plugin

* feat(pwa) make site work offline and respect SSG css

* feat(pwa) offline works for home and for other pages

* feat(pwa) use proper app shell for all pages cept home

* feat(pwa) make real app shell for doc pages, only pre cache hp and app shell, now everything offline
